fix: use int instead of int64 for EventID

+++ observe_events_options.go
@@ -3,7 +3,7 @@ package eventsourcingdb
 type ObserveEventsOptions struct {
 	OptionWithSubStreams *bool     `json:"withSubStreams,omitempty"`
 	OptionEventNames     *[]string `json:"eventNames,omitempty"`
-	OptionLowerBoundID   *int64    `json:"lowerBoundId,omitempty"`
+	OptionLowerBoundID   *int      `json:"lowerBoundId,omitempty"`
 	OptionFromEventName  *string   `json:"fromEventName,omitempty"`
 }
 
@@ -23,7 +23,7 @@ func (options ObserveEventsOptions) EventNames(eventNames []string) ObserveEvent
 	return options
 }
 
-func (options ObserveEventsOptions) LowerBoundID(lowerBoundID int64) ObserveEventsOptions {
+func (options ObserveEventsOptions) LowerBoundID(lowerBoundID int) ObserveEventsOptions {
 	options.OptionLowerBoundID = &lowerBoundID
 
 	return options
